requirements.txt
Flask==1.1.1
Flask-Sockets==0.2.1
gevent==1.4.0
gevent-websocket==0.10.1
greenlet==0.4.15
google-cloud-speech
websocket-client